projects-2026/knowledge-seeder/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-seeder"
version = "0.1.0"
description = "CLI tool for batch ingestion of knowledge sources into Knowledge Engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Knowledge Seeder Team" }
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.20.0",
    "trafilatura>=1.12.0",
    "youtube-transcript-api>=0.6.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
knowledge-seeder = "knowledge_seeder.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_seeder"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
